// scripts/types/generated/mock-red-bank/MockRedBank.types.ts
// @ts-nocheck
/**
* This file was automatically generated by @cosmwasm/ts-codegen@0.16.1.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export type Decimal = string;
export interface InstantiateMsg {
  coins: CoinMarketInfo[];
  [k: string]: unknown;
}
export interface CoinMarketInfo {
  denom: string;
  liquidation_threshold: Decimal;
  max_ltv: Decimal;
  [k: string]: unknown;
}
export type ExecuteMsg = {
  borrow: {
    coin: Coin;
    recipient?: string | null;
    [k: string]: unknown;
  };
} | {
  repay: {
    denom: string;
    on_behalf_of?: string | null;
    [k: string]: unknown;
  };
};
export type Uint128 = string;
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
export type QueryMsg = {
  user_asset_debt: {
    denom: string;
    user_address: string;
  };
} | {
  market: {
    denom: string;
  };
};
export interface Market {
  borrow_enabled: boolean;
  borrow_index: Decimal;
  borrow_rate: Decimal;
  debt_total_scaled: Uint128;
  denom: string;
  deposit_cap: Uint128;
  deposit_enabled: boolean;
  indexes_last_updated: number;
  liquidation_bonus: Decimal;
  liquidation_threshold: Decimal;
  liquidity_index: Decimal;
  liquidity_rate: Decimal;
  max_loan_to_value: Decimal;
  reserve_factor: Decimal;
  [k: string]: unknown;
}
export interface UserAssetDebtResponse {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}